% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{blockMax}
\alias{blockMax}
\title{Compute block maxima of a vector}
\usage{
blockMax(x, blockSize, overlap = 0, type = "C++", na.rm = FALSE)
}
\arguments{
\item{x}{Vector that is split into blocks}

\item{blockSize}{Integer, size of each block.}

\item{overlap}{Integer, number of overlapping data points.}

\item{type}{Toggle between C++ ("C++") and R ("R"), see chunk.}

\item{na.rm}{Boolean, is passed on to max}
}
\value{
Returns List, with the maximum per block (blockMax) and an indicator if chunk has produced NaN values. Then the index of the first NaN (firstNaN) is returned. This happens if the blockSize does not divide the length of x without remainder; see chunk.
}
\description{
Compute block maxima of a vector
}
