% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{blockMaxCpp}
\alias{blockMaxCpp}
\title{Compute block maxima of a vector via C++}
\usage{
blockMaxCpp(x, blockSize, overlap = 0, type = "C++", na.rm = FALSE)
}
\arguments{
\item{x}{Vector that is split into blocks}

\item{blockSize}{Integer, size of each block.}

\item{overlap}{Integer, number of overlapping data points.}

\item{type}{Toggle between C++ ("C++") and R ("R"), see chunk.}

\item{na.rm}{Boolean, is passed on to max}
}
\value{
Returns List, with the maximum per block (blockMax) and an indicator if chunk has produced NaN values. Then the index of the first NaN (firstNaN) is returned. This happens if the blockSize does not divide the length of x without remainder; see chunk. If there is no NaN value the indicator is zero. For the computation all NA and NaN values are ignored.
}
\description{
Compute block maxima of a vector via C++
}
