% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.simulation.plan.R
\name{check.computation.plan}
\alias{check.computation.plan}
\title{Check which computations of a compute plan are done}
\usage{
check.computation.plan(
  strPlan,
  zipDir = ".",
  pattern = "*_out.zip",
  hashFile = "inputFile.hash"
)
}
\arguments{
\item{strPlan}{A string specifying the RData file with the compute plan.}

\item{zipDir}{A string specifying the folder with the computation results.}

\item{pattern}{A string to identify the file inside the zip folder that contains the hash connected to the starting value file (a row in the compute plan).}
}
\description{
Check which computations of a compute plan are done
}
