% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk.NaN}
\alias{chunk.NaN}
\title{Return first line with NaN when using chunk}
\usage{
chunk.NaN(x, chunkSize, overlap = 0)
}
\arguments{
\item{x}{Length of input vector for chunk}

\item{chunkSize}{Integer, size of each block.}

\item{overlap}{Integer, number of overlapping data points.}

\item{Returns}{0 if no NaN values are imputed and an integer line number with the first imputed NaN value otherwise.}
}
\description{
Return first line with NaN when using chunk
}
