% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Partition vector into blocks (possibly overlapping)}
\usage{
chunk(x, chunkSize, overlap = 0, type = "C++")
}
\arguments{
\item{x}{Input vector}

\item{chunkSize}{Integer, size of each block.}

\item{overlap}{Integer, number of overlapping data points.}

\item{type}{Toggle either the R or the C++ implementation.}
}
\description{
Partition vector into blocks (possibly overlapping)
}
