% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunkR}
\alias{chunkR}
\title{Partition vector into blocks (possibly overlapping) - R implementation}
\usage{
chunkR(x, chunkSize, overlap = 0)
}
\arguments{
\item{x}{Input vector}

\item{chunkSize}{Integer, size of each block.}

\item{overlap}{Integer, number of overlapping data points.}
}
\description{
Partition vector into blocks (possibly overlapping) - R implementation
}
