% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{citeName}
\alias{citeName}
\title{Function that transforms First Middle and Last name into a standard citation style string}
\usage{
citeName(first, middle, last)
}
\arguments{
\item{first}{First name of the person}

\item{middle}{Middle name of the person (or NA)}

\item{last}{Last name of the person}
}
\value{
A string with the last name followed by the extracted initials
}
\description{
The function transforms names into standard citation style strings. For example Rainer Maria Rilke is converted to Rilke, R.M.
The function breaks at spaces or hyphen, see the examples.
}
\examples{
print(citeName("Rainer","Maria","Rilke")) #Rilke, R.M.
print(citeName("Jean-Luc", "Borg", "Picard")) #Picard, J.L.B.
}
\author{
Klaus Herrmann
}
