% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{conferenceDetails}
\alias{conferenceDetails}
\title{Extract the information linked to a conference venue.}
\usage{
conferenceDetails(xmlNode, strLanguage = ".//eng")
}
\arguments{
\item{xmlNode}{The xml node with the information}

\item{strLanguage}{A string specifying which language should be used}
}
\value{
A list containing the information (event, location, misc) extracted from the xml node
}
\description{
The function extracts the event, location and miscellaneous information from a xml node linked to a poster or a talk.
This gives the event name, the location (it is necessary to have a sub-node in each language: Brussels vs. Bruxelles) and miscellaneous information in a list.
}
\author{
Klaus Herrmann
}
