% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatInCollectionMD}
\alias{formatInCollectionMD}
\title{Convert in collection list to a markdown string}
\usage{
formatInCollectionMD(lstInCollection)
}
\arguments{
\item{lstInCollection}{The list with the information}
}
\value{
A string in the markdown format with a standard citation for the article
}
\description{
The function transforms a list representing an in collection publication into a markdown string.
}
\author{
Klaus Herrmann
}
