% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatJournalArticleMD}
\alias{formatJournalArticleMD}
\title{Convert journal article list to a markdown string}
\usage{
formatJournalArticleMD(lstArticle)
}
\arguments{
\item{lstArticle}{The list with the information}
}
\value{
A string in the markdown format with a standard citation for the article
}
\description{
The function transforms a list representing a journal article into a markdown string.
}
\author{
Klaus Herrmann
}
