% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatOtherMD}
\alias{formatOtherMD}
\title{Convert other publication list to a markdown string}
\usage{
formatOtherMD(lstOther)
}
\arguments{
\item{lstArticle}{The list with the information}
}
\value{
A string in the markdown format with a standard citation for the article (other)
}
\description{
The function transforms a list representing a other publication into a markdown string.
}
\author{
Klaus Herrmann
}
