% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatPosterMD}
\alias{formatPosterMD}
\title{Convert poster list to a markdown string}
\usage{
formatPosterMD(lstPoster, strSetlocale = c("LC_ALL", "en_US.UTF-8"))
}
\arguments{
\item{lstPoster}{The list with the information}

\item{strSetlocale}{The list with the information}
}
\value{
A string in the markdown format with a standard citation for the poster
}
\description{
The function transforms a list representing a poster into a markdown string.
The second parameter allows to change the language used for the months (April vs. avril).
Possible values are c(LC_ALL","en_US.UTF-8") or c(LC_ALL","fr_FR.UTF-8"). German is somehow not working.
}
\author{
Klaus Herrmann
}
