% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatPresentationMD}
\alias{formatPresentationMD}
\title{Convert presentation (talk) list to a markdown string}
\usage{
formatPresentationMD(
  lstPresentation,
  strSetlocale = c("LC_ALL", "en_US.UTF-8")
)
}
\arguments{
\item{lstPresentation}{The list with the information}

\item{strSetlocale}{The list with the information}
}
\value{
A string in the markdown format with a standard citation for the talk
}
\description{
The function transforms a list representing a presentation (talk) into a markdown string.
The second parameter allows to change the language used for the months (April vs. avril).
Possible values are c(LC_ALL","en_US.UTF-8") or c(LC_ALL","fr_FR.UTF-8"). German is somehow not working.
}
\author{
Klaus Herrmann
}
