% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{formatTeachingMD}
\alias{formatTeachingMD}
\title{Convert course (teaching) list to a markdown string}
\usage{
formatTeachingMD(lstCours)
}
\arguments{
\item{lstCours}{The list with the information}
}
\value{
A string in the markdown format with a representation of the cours
}
\description{
The function transforms a list representing a course (teaching) into a markdown string.
}
\author{
Klaus Herrmann
}
