% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booleanCheckFunctions.R
\name{is.complexNumber}
\alias{is.complexNumber}
\title{Check if an argument is a complex scalar value.}
\usage{
is.complexNumber(x)
}
\arguments{
\item{x}{Value to be tested.}
}
\value{
Boolean TRUE if it is a complex scalar value, FALSE otherwise
}
\description{
Check if an argument is a complex scalar value.
}
