% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booleanCheckFunctions.R
\name{is.integerNumber}
\alias{is.integerNumber}
\title{Check if an argument is a scalar integer.}
\usage{
is.integerNumber(x)
}
\arguments{
\item{x}{Value to be tested.}
}
\value{
Boolean TRUE if it is a integer scalar value, FALSE otherwise
}
\description{
Check if an argument is a scalar integer.
}
