% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/booleanCheckFunctions.R
\name{is.realNumber}
\alias{is.realNumber}
\title{Check if an argument is a real scalar value.}
\usage{
is.realNumber(x)
}
\arguments{
\item{x}{Value to be tested.}
}
\value{
Boolean TRUE if it is a real scalar value, FALSE otherwise
}
\description{
Check if an argument is a real scalar value.
}
