% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.with.summary.statistic.R
\name{khBoxplot}
\alias{khBoxplot}
\title{Boxplot that adds an additional summary statistic (default: mean)}
\usage{
khBoxplot(
  x,
  summaryFunc = mean,
  mpch = 19,
  mcol = "black",
  addSummary = TRUE,
  ...
)
}
\arguments{
\item{x}{List, data.frame or matrix for where for each column a boxplot is made}

\item{summaryFunc}{A function that computes a summary statistic for a vector of values. Default is mean}

\item{mpch}{The pch value used to plot the value of summaryFunc in the plot}

\item{mcol}{Color (or vector of colors) used to plot the value of summaryFunc}

\item{addSummary}{Boolean toggle to plot the summaryFunc or not.}

\item{...}{Parameters passed to boxplot}
}
\value{
A boxplot with the value of the summary statistic indicated.
}
\description{
Boxplot that adds an additional summary statistic (default: mean)
}
