% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khermisc.R
\docType{khermisc}
\name{khermisc}
\alias{khermisc}
\title{khermisc: A package collecting different useful functions without a specific purpose.}
\description{
The khermisc package collects functions without a specific purpose. The functions can be in different categories:
Simple utility functions or to help with computations on clusters.
}
\section{khermisc functions}{

The mypackage functions ...
}

