% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OkabeIto.R
\name{OkabeIto}
\alias{OkabeIto}
\title{Get a color blind friendly color palette with eight colors}
\usage{
OkabeIto(x = 1:8)
}
\arguments{
\item{x}{order of the palette}
}
\value{
vector with colors in the specified order
}
\description{
Get a color blind friendly color palette with eight colors
}
