% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{personsDataFrame}
\alias{personsDataFrame}
\title{Extract the people linked to a publication. This can be the authors or the editors.}
\usage{
personsDataFrame(xmlNode, strCategory = ".//author")
}
\arguments{
\item{xmlNode}{A xml node representing one article}

\item{strCategory}{A string indicating which sub-elements should be used}
}
\value{
A data.frame containing sorted by the indicated ordering, containing the first middle and last name for each person
}
\description{
When an xml file representing publications is given, the function extracts the authors (and editors) into an R list.
}
\author{
Klaus Herrmann
}
