% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.simulation.plan.R
\name{row.hash}
\alias{row.hash}
\title{Compute hash value from a vector by collapsing it into one string}
\usage{
row.hash(vec)
}
\arguments{
\item{vec}{A vector for which the hash is computed.}
}
\description{
Compute hash value from a vector by collapsing it into one string
}
