% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.job.R
\name{slurm.job.R}
\alias{slurm.job.R}
\title{Builds a shell script that runs a job on a compute cluster}
\usage{
slurm.job.R(
  maxTime,
  account,
  memory,
  modules,
  exports,
  scriptName,
  inputFile,
  inputFolder,
  outputFolder = ""
)
}
\arguments{
\item{maxTime}{A string specifying the maximum run time after which the job is canceled.}

\item{account}{A string specifying the account for billing.}

\item{memory}{A string specifying the memory requirements for the job.}

\item{modules}{A vector of strings with the modules.}

\item{exports}{A vector of strings with exports.}

\item{scriptName}{A string that identifies where the R script is located (full path).}

\item{inputFile}{A string that identifies the input RData file (file name only).}

\item{inputFolder}{A string that identifies the folder where the inputFile is located.}

\item{outputFolder}{A string that identifies where the output file is located.}
}
\description{
Builds a shell script that runs a job on a compute cluster
}
