% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.job.R
\name{slurm.job.standard.exports}
\alias{slurm.job.standard.exports}
\title{Returns standard exports that are used for R on the Compute Canada cluster}
\usage{
slurm.job.standard.exports()
}
\value{
A vector of strings that can be used as input to for the exports argument in slurm.job.R
}
\description{
Returns standard exports that are used for R on the Compute Canada cluster
}
