% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.job.R
\name{slurm.job.standard.modules}
\alias{slurm.job.standard.modules}
\title{Returns standard modules that are used for R on the Compute Canada cluster}
\usage{
slurm.job.standard.modules()
}
\value{
A vector of strings that can be used as input to for the modules argument in slurm.job.R
}
\description{
Returns standard modules that are used for R on the Compute Canada cluster
}
