% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.job.R
\name{slurm.job.standard.run}
\alias{slurm.job.standard.run}
\title{Returns standard shell code that can be used for R on the Compute Canada cluster}
\usage{
slurm.job.standard.run(CompPlanName, row, scriptName)
}
\arguments{
\item{row}{An integer specifying the row of the compute plan that should be used for the computations.}

\item{scriptName}{A string specifying the name of the R script that will perform the calculations.}

\item{inputFileName}{A string specifying the name of the compute plan file without the file ending .CompPlan}
}
\value{
A list with a vector of strings that can be used as input to for the strRun argument in slurm.job.R, and a string that is the name of the shell script that one should use to write it.
}
\description{
Returns standard shell code that can be used for R on the Compute Canada cluster
}
