% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.simulation.plan.R
\name{to.string}
\alias{to.string}
\title{Convert object to string if print() is overloaded}
\usage{
to.string(obj)
}
\arguments{
\item{obj}{An object for which print() returns a vector of strings}
}
\value{
A string representing the object.
}
\description{
Convert object to string if print() is overloaded
}
