% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{xml2lstPoster}
\alias{xml2lstPoster}
\title{Create an R list object from a xml Poster node}
\usage{
xml2lstPoster(xmlPoster, dateFormat = "\%Y/\%m/\%d", strLanguage = ".//eng")
}
\arguments{
\item{xmlPoster}{The xml node with the information}

\item{dateFormat}{A string specifying how the dates in the xml node are represented}

\item{strLanguage}{A string specifying which language should be used}
}
\value{
A list containing the information extracted from the xml node
}
\description{
The function transforms a xml node representing a poster presentation into a list.
}
\author{
Klaus Herrmann
}
