% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{xml2lstPresentation}
\alias{xml2lstPresentation}
\title{Create an R list object from a xml Presentation (Talk) node}
\usage{
xml2lstPresentation(
  xmlPresentation,
  dateFormat = "\%Y/\%m/\%d",
  strLanguage = ".//eng"
)
}
\arguments{
\item{dateFormat}{A string specifying how the dates in the xml node are represented}

\item{strLanguage}{A string specifying which language should be used}

\item{xmlPoster}{The xml node with the information}
}
\value{
A list containing the information extracted from the xml node
}
\description{
The function transforms a xml node representing a talk into a list.
}
\author{
Klaus Herrmann
}
