% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV_Tools.R
\name{xml2lstTeaching}
\alias{xml2lstTeaching}
\title{Create an R list object from a xml node representing a course (teaching)}
\usage{
xml2lstTeaching(xmlNode, dateFormat = "\%Y-\%m-\%d")
}
\arguments{
\item{xmlNode}{The xml node with the information}

\item{dateFormat}{A string specifying how the dates in the xml node are represented}
}
\value{
A list containing the information extracted from the xml node
}
\description{
The function transforms a xml node representing a course into a list.
}
\author{
Klaus Herrmann
}
