% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.frame.to.latex.table.string.R
\name{data.frame.to.latex.table.string}
\alias{data.frame.to.latex.table.string}
\title{Output the contents of a numeric matrix or dataframe into LaTeX type table notation string}
\usage{
data.frame.to.latex.table.string(
  dframe,
  topLeft = "",
  fmt = "\%f",
  header = TRUE
)
}
\arguments{
\item{dframe}{A data.frame or matrix}

\item{topLeft}{A string that will fill the top left corner (first cell of the first row) of the table}

\item{fmt}{A parameter that gets passed to sprintf in case dframe is a numeric matrix}

\item{header}{A boolean toggle. If true a full latex table with \begin{table} etc is created, if false only cells are converted.}
}
\value{
A vector of strings where each entry is one line of latex code.
}
\description{
Output the contents of a numeric matrix or dataframe into LaTeX type table notation string
}
\examples{
n <- 3
m <- 6
M1 <- matrix(1:(n*m),n,m)
rownames(M1) <- 1:n
colnames(M1) <- 1:m
S1 <- data.frame.to.latex.table.string(M1,topLeft="TL")
}
