% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.job.R
\name{slurm.job.R}
\alias{slurm.job.R}
\title{Builds a shell script that runs a job on a compute cluster}
\usage{
slurm.job.R(
  maxTime,
  account,
  memory,
  modules = slurm.job.standard.modules(),
  exports = slurm.job.standard.exports(),
  strRun
)
}
\arguments{
\item{maxTime}{A string specifying the maximum run time after which the job is canceled.}

\item{account}{A string specifying the account for billing.}

\item{memory}{A string specifying the memory requirements for the job.}

\item{modules}{A vector of strings with the modules.}

\item{exports}{A vector of strings with exports.}

\item{strRun}{A vector of strings with the instructions (bash code) for the job.}
}
\value{
A string that can be written with writeLines to a .sh file that can be executed on a cluster.
}
\description{
Builds a shell script that runs a job on a compute cluster
}
